+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.4.0
+
+### Sentence corpus creation
+
+- ALTO input now uses block-level tags for filtering and section type in sentence corpus
+    - By default, only includes blocks tagged as text, footnote, Title, or untagged
+
 ## [0.3.0] - 2025-10-27
 
 ### Sentence corpus creation